+++ wpull/testing/badapp.py
@@ -33,6 +33,7 @@ class Handler(http.server.BaseHTTPRequestHandler):
             '/content_length': self.basic_content_length,
             '/chunked': self.basic_chunked,
             '/chunked_trailer': self.basic_chunked_trailer,
+            '/chunked_trailer_2': self.basic_chunked_trailer_2,
             '/chunked_non_standard_delim': self.chunked_non_standard_delim,
             '/chunked_with_extension': self.chunked_with_extension,
             '/underrun': self.underrun_response,
@@ -111,6 +112,13 @@ class Handler(http.server.BaseHTTPRequestHandler):
         self.end_headers()
         self.wfile.write(
             b'5\r\nhello\r\n0007\r\n world!\r\n0\r\nAnimal: dolphin\r\n\r\n')
+
+    def basic_chunked_trailer_2(self):
+        self.send_response(200)
+        self.send_header('transfer-encoding', 'chunked')
+        self.end_headers()
+        self.wfile.write(b'5\r\nhello\r\n0007\r\n world!\r\n0\r\n')
+        self.wfile.write(b'Animal: dolphin\r\nCake: delicious\r\n\r\n')
 
     def chunked_non_standard_delim(self):
         self.send_response(200)